+++ CHANGELOG.md
@@ -20,3 +20,5 @@ and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec
 - Add editorconfig
 - Remove example sas token
 - Remove redundant modules
+- Azure VM module for azure windows images
+- Azure VM module for custom windows images